Create README.md
For my project analysis, I’ll use python’s libraries. For fetching the information from the dataset I’ll use pandas library and pd.read_csv() method. And with help of pandas’ other methods we can get the insight on the GTD dataset. In case of any null or irrelevant data, cleaning of data is required with help of mean, median and mode.